graphql-server/requirements.txt
graphql-core>=3
Mako==1.0.13